Code refactoring complete
